requirements.txt
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
autopep8==1.6.0
